#25 recap ok

+++ src/main/java/com/kifiya/kobiri/repositories/GerantRepository.java
@@ -74,7 +74,7 @@ public class GerantRepository {
             transfert.setDateTransfert(resultSet.getTimestamp("date_transfert").toLocalDateTime());
             transfert.setDateValidation(resultSet.getTimestamp("date_validation").toLocalDateTime());
             transfert.setMontantEuros(resultSet.getLong("montant_euros"));
-            transfert.setTaux(resultSet.getLong("taux"));
+            transfert.setTaux(resultSet.getDouble("taux"));
             return transfert;
         });
     }
